server/microservices/requirements.txt
gcsfs
google-auth
PyJWT
pandas
dacite
requests
cryptography
protobuf==3.20.*
gtfs-realtime-bindings
google-cloud-secret-manager